Putting in Some UI - Adding Fitness and Fitness based genepool - Trying only allowing higher average fitness generations

+++ src/js/genetics/population.js
@@ -7,11 +7,16 @@ import * as tf from "@tensorflow/tfjs";
  * and the necessary selection and crossbreeding functions
  */
 class Population {
-  constructor(size, config, mutation) {
+  constructor(size, config, mutation, allowLowerFitness = true) {
     this.size = size;
     this.config = config;
     this.year = 0;
     this.mutation = mutation;
+    this.avgFitness = 0;
+    this.genePool = [];
+    this.oldGenePool = [];
+    this.stuckCicles = 0;
+    this.lowFitness = allowLowerFitness;
     this.init();
   }
 
@@ -30,6 +35,8 @@ class Population {
    */
   setupGames() {
     this.games = [];
+    this.oldGenePool = this.genePool.slice(0);
+    this.genePool = [];
     shuffle(this.population);
     for (let i = 0; i < this.size / 2; i++) {
       this.games.push({
@@ -56,20 +63,62 @@ class Population {
    */
   breed() {
     this.year++;
+    this.avgFitness = this.genePool.length / 10.0;
+    const genFailed =
+      this.avgFitness < this.oldGenePool.length / 10.0 || this.avgFitness >= 9;
+    //NoLowFitnessAllowed Mechanism
+    if (!this.lowFitness) {
+      if (genFailed && this.stuckCicles < 100) {
+        this.stuckCicles++;
+        this.genePool.forEach(element => {
+          if (
+            !this.oldGenePool.includes(element) &&
+            !this.nextPopulation.includes(element)
+          ) {
+            try {
+              element.kill();
+            } catch (error) {
+              console.log("Already disposed");
+            }
+          }
+        });
+        this.genePool = this.oldGenePool;
+      } else {
+        this.stuckCicles = 0;
+        this.oldGenePool.forEach(element => {
+          if (!this.genePool.includes(element)) {
+            try {
+              element.kill();
+            } catch (error) {
+              console.log("Already disposed");
+            }
+          }
+        });
+        console.log(tf.memory());
+      }
+    }
     tf.tidy(() => {
-      const parents = this.nextPopulation.slice(0);
-      for (let k = 0; k < 2; k++) {
-        shuffle(parents);
-        for (let i = 0; i < this.size / 4; i++) {
-          this.nextPopulation.push(
-            this.cross(parents[i * 2], parents[i * 2 + 1])
-          );
-        }
+      let mom = this.genePool[0];
+      let dad;
+      do {
+        shuffle(this.genePool);
+        dad = this.genePool[0];
+      } while (mom == dad);
+
+      for (let i = 0; i < this.size / 2; i++) {
+        this.nextPopulation.push(this.cross(mom, dad));
       }
     });
 
     this.mutate();
     this.population = this.nextPopulation;
+
+    //Recursive until not failing anymore
+    if (genFailed) {
+      this.setupGames();
+      this.play();
+      this.breed();
+    }
   }
 
   /**
@@ -163,27 +212,35 @@ class Population {
         switch (currentplayer) {
           case 0:
             this.nextPopulation.push(game.p2);
-            game.p1.kill();
+            this.genePool.push(
+              ...new Array(game.p2.fitness(game.game)).fill(game.p2)
+            );
+            this.lowFitness && game.p1.kill();
             break;
           case 1:
             this.nextPopulation.push(game.p1);
-            game.p2.kill();
+            this.genePool.push(
+              ...new Array(game.p1.fitness(game.game)).fill(game.p1)
+            );
+            this.lowFitness && game.p2.kill();
             break;
           default:
             break;
         }
         game.finished = true;
       } else if (winner !== 0) {
-        console.log("A Fucking Bot won with " + gameResult.row);
+        console.log("A Bot won with " + gameResult.row);
         console.log(game.game.getStandings());
         switch (winner - 1) {
           case 0:
             this.nextPopulation.push(game.p1);
-            game.p2.kill();
+            this.genePool.push(...new Array(8).fill(game.p1));
+            this.lowFitness && game.p2.kill();
             break;
           case 1:
             this.nextPopulation.push(game.p2);
-            game.p1.kill();
+            this.genePool.push(...new Array(8).fill(game.p2));
+            this.lowFitness && game.p1.kill();
             break;
           default:
             break;
